[metasrv] refactor: move RPC message trait impls to message.rs
